requirements.txt
streamlit>=1.28.0
pandas
geopy
python-dateutil
Pillow
pytesseract
pdf2image
pytest
